pyproject.toml
[tool.poetry]
name = "book_library"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.1.3"
djangorestframework = "^3.14.0"
pydantic = "^1.10.2"
dependency-injector = "^4.40.0"
factory-boy = "^3.2.1"
djangorestframework-simplejwt = "^5.2.2"
django-extensions = "^3.2.1"
requests = "^2.28.2"
black = "^23.3.0"
isort = "^5.12.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
